+++ src/sora_client_context.cpp
@@ -1,4 +1,4 @@
-#include "sora/sora_default_client.h"
+#include "sora/sora_client_context.h"
 
 // WebRTC
 #include <api/create_peerconnection_factory.h>
@@ -14,10 +14,6 @@
 #include <rtc_base/logging.h>
 #include <rtc_base/ssl_adapter.h>
 
-#ifdef _WIN32
-#include <rtc_base/win/scoped_com_initializer.h>
-#endif
-
 #include "sora/audio_device_module.h"
 #include "sora/camera_device_capturer.h"
 #include "sora/java_context.h"
@@ -29,23 +25,35 @@
 
 namespace sora {
 
-SoraDefaultClient::SoraDefaultClient(SoraDefaultClientConfig config)
-    : config_(config) {}
+SoraClientContext::~SoraClientContext() {
+  config_ = SoraClientContextConfig();
+  connection_context_ = nullptr;
+  factory_ = nullptr;
+  network_thread_->Stop();
+  worker_thread_->Stop();
+  signaling_thread_->Stop();
 
-bool SoraDefaultClient::Configure() {
+  //rtc::CleanupSSL();
+}
+
+std::shared_ptr<SoraClientContext> SoraClientContext::Create(
+    const SoraClientContextConfig& config) {
   rtc::InitializeSSL();
 
-  network_thread_ = rtc::Thread::CreateWithSocketServer();
-  network_thread_->Start();
-  worker_thread_ = rtc::Thread::Create();
-  worker_thread_->Start();
-  signaling_thread_ = rtc::Thread::Create();
-  signaling_thread_->Start();
+  std::shared_ptr<SoraClientContext> c = std::make_shared<SoraClientContext>();
+
+  c->config_ = config;
+  c->network_thread_ = rtc::Thread::CreateWithSocketServer();
+  c->network_thread_->Start();
+  c->worker_thread_ = rtc::Thread::Create();
+  c->worker_thread_->Start();
+  c->signaling_thread_ = rtc::Thread::Create();
+  c->signaling_thread_->Start();
 
   webrtc::PeerConnectionFactoryDependencies dependencies;
-  dependencies.network_thread = network_thread_.get();
-  dependencies.worker_thread = worker_thread_.get();
-  dependencies.signaling_thread = signaling_thread_.get();
+  dependencies.network_thread = c->network_thread_.get();
+  dependencies.worker_thread = c->worker_thread_.get();
+  dependencies.signaling_thread = c->signaling_thread_.get();
   dependencies.task_queue_factory = webrtc::CreateDefaultTaskQueueFactory();
   dependencies.call_factory = webrtc::CreateCallFactory();
   dependencies.event_log_factory =
@@ -57,14 +65,17 @@ bool SoraDefaultClient::Configure() {
   // media_dependencies
   cricket::MediaEngineDependencies media_dependencies;
   media_dependencies.task_queue_factory = dependencies.task_queue_factory.get();
-  media_dependencies.adm = worker_thread_->BlockingCall([&] {
+  media_dependencies.adm = c->worker_thread_->BlockingCall([&] {
     sora::AudioDeviceModuleConfig config;
-    if (!config_.use_audio_device) {
+    if (!c->config_.use_audio_device) {
       config.audio_layer = webrtc::AudioDeviceModule::kDummyAudio;
     }
     config.task_queue_factory = dependencies.task_queue_factory.get();
     config.jni_env = sora::GetJNIEnv();
-    config.application_context = GetAndroidApplicationContext(config.jni_env);
+    if (c->config_.get_android_application_context) {
+      config.application_context =
+          c->config_.get_android_application_context(config.jni_env);
+    }
     return sora::CreateAudioDeviceModule(config);
   });
 
@@ -76,7 +87,7 @@ bool SoraDefaultClient::Configure() {
   auto cuda_context = sora::CudaContext::Create();
   {
     auto config =
-        config_.use_hardware_encoder
+        c->config_.use_hardware_encoder
             ? sora::GetDefaultVideoEncoderFactoryConfig(cuda_context, env)
             : sora::GetSoftwareOnlyVideoEncoderFactoryConfig();
     config.use_simulcast_adapter = true;
@@ -85,7 +96,7 @@ bool SoraDefaultClient::Configure() {
   }
   {
     auto config =
-        config_.use_hardware_encoder
+        c->config_.use_hardware_encoder
             ? sora::GetDefaultVideoDecoderFactoryConfig(cuda_context, env)
             : sora::GetSoftwareOnlyVideoDecoderFactoryConfig();
     media_dependencies.video_decoder_factory =
@@ -96,28 +107,32 @@ bool SoraDefaultClient::Configure() {
   media_dependencies.audio_processing =
       webrtc::AudioProcessingBuilder().Create();
 
+  if (c->config_.configure_media_dependencies) {
+    c->config_.configure_media_dependencies(dependencies, media_dependencies);
+  }
+
   dependencies.media_engine =
       cricket::CreateMediaEngine(std::move(media_dependencies));
 
-  ConfigureDependencies(dependencies);
+  if (c->config_.configure_dependencies) {
+    c->config_.configure_dependencies(dependencies);
+  }
 
-  factory_ = sora::CreateModularPeerConnectionFactoryWithContext(
-      std::move(dependencies), connection_context_);
+  c->factory_ = sora::CreateModularPeerConnectionFactoryWithContext(
+      std::move(dependencies), c->connection_context_);
 
-  if (factory_ == nullptr) {
+  if (c->factory_ == nullptr) {
     RTC_LOG(LS_ERROR) << "Failed to create PeerConnectionFactory";
-    return false;
+    return nullptr;
   }
 
   webrtc::PeerConnectionFactoryInterface::Options factory_options;
   factory_options.disable_encryption = false;
   factory_options.ssl_max_version = rtc::SSL_PROTOCOL_DTLS_12;
   factory_options.crypto_options.srtp.enable_gcm_crypto_suites = true;
-  factory_->SetOptions(factory_options);
+  c->factory_->SetOptions(factory_options);
 
-  OnConfigured();
-
-  return true;
+  return c;
 }
 
 }  // namespace sora